Move public to src/assets
