+++ src/assets/fonts/interui/inter-ui-bolditalic.woff
