+++ src/assets/icons/index.js
