+++ src/base/variables/_paths.scss
@@ -4,6 +4,6 @@
 // *****************************
 //------------------------------
 
-$path-fonts: "../../public/fonts/" !default;
-$path-images: "../../public/images/" !default;
-$path-icons: "../../public/icons/" !default;
+$path-fonts: "../assets/fonts/" !default;
+$path-images: "../assets/images/" !default;
+$path-icons: "../assets/icons/" !default;
